--- a/.gitignore
+++ b/.gitignore
@@ -23,3 +23,4 @@ out.log
 *secrets
 *.csv
 */vimeo/*.txt
+api/logs/