--- a/YT8M_downloader/downloader.py
+++ b/YT8M_downloader/downloader.py
@@ -12,7 +12,7 @@ import youtube_dl
 
 parser = argparse.ArgumentParser(description='Download dataset')
 parser.add_argument('-o', '--output', action='store', help='Folder where the videos will be', type=str, required=True)
-parser.add_argument('-f', '--format', action='store', help='YT downloader video format filter. 137 for 1080p30fps, 299 for 1080p60fps.', type=str, required=False, default='137')
+parser.add_argument('-f', '--format', action='store', help='YT downloader video format filter. 137 for 1080p30fps, 299 for 1080p60fps. For more info, see youtube_dl package documentation.', type=str, required=False, default='137')
 
 args = parser.parse_args()
 format_filter = args.format