--- a/feature_engineering/notebooks/Fps_renditions.ipynb
+++ b/feature_engineering/notebooks/Fps_renditions.ipynb
@@ -2,16 +2,15 @@
  "cells": [
   {
    "cell_type": "markdown",
-   "source": [
-    "# Classification metrics for FPS-adjusted renditions\n",
-    "#### Data is generated with testing/test_var_fps_classification.py script"
-   ],
    "metadata": {
-    "collapsed": false,
     "pycharm": {
      "name": "#%% md\n"
     }
-   }
+   },
+   "source": [
+    "# Classification metrics for FPS-adjusted renditions\n",
+    "#### Data is generated with testing/test_var_fps_classification.py script"
+   ]
   },
   {
    "cell_type": "code",
@@ -50,7 +49,7 @@
    "outputs": [],
    "source": [
     "# estimate classification metrics\n",
-    "df['prediction'] = df['score']>0"
+    "df['prediction'] = df['score']<0"
    ]
   },
   {