--- a/feature_engineering/notebooks/Fps_renditions.ipynb
+++ b/feature_engineering/notebooks/Fps_renditions.ipynb
@@ -1083,9 +1083,7 @@
      "output_type": "execute_result"
     }
    ],
-   "source": [
-    "df"
-   ]
+   "source": []
   },
   {
    "cell_type": "markdown",