--- a/feature_engineering/notebooks/Fps_renditions.ipynb
+++ b/feature_engineering/notebooks/Fps_renditions.ipynb
@@ -14,18 +14,21 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 17,
+   "execution_count": 94,
    "metadata": {},
    "outputs": [],
    "source": [
     "import pandas as pd\n",
     "import numpy as np\n",
-    "from sklearn.metrics import classification_report"
+    "from sklearn.metrics import classification_report\n",
+    "pd.options.display.width = 0\n",
+    "pd.set_option('display.max_columns', None)\n",
+    "from IPython.display import display"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 18,
+   "execution_count": 95,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -41,7 +44,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 19,
+   "execution_count": 96,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -51,13 +54,12 @@
    "source": [
     "# estimate classification metrics\n",
     "df['ground_truth'] = np.where(df['is_correct'], 'Correct', 'Tamper')\n",
-    "df['prediction'] = np.where(df['score']>0, 'Correct', 'Tamper')\n",
-    "df['prediction_old'] = np.where(df['score_old']>0, 'Correct', 'Tamper')"
+    "df['prediction'] = np.where(df['score']>0, 'Correct', 'Tamper')"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 21,
+   "execution_count": 97,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -70,21 +72,10 @@
      "output_type": "stream",
      "text": [
       "Non-framerate adjusted renditions:\n",
-      "Timestamp-based frame selection:\n",
       "              precision    recall  f1-score   support\n",
       "\n",
       "     Correct       0.74      0.94      0.83        64\n",
       "      Tamper       0.91      0.67      0.77        64\n",
-      "\n",
-      "    accuracy                           0.80       128\n",
-      "   macro avg       0.83      0.80      0.80       128\n",
-      "weighted avg       0.83      0.80      0.80       128\n",
-      "\n",
-      "Index-based frame selection:\n",
-      "              precision    recall  f1-score   support\n",
-      "\n",
-      "     Correct       0.73      0.95      0.83        64\n",
-      "      Tamper       0.93      0.66      0.77        64\n",
       "\n",
       "    accuracy                           0.80       128\n",
       "   macro avg       0.83      0.80      0.80       128\n",
@@ -96,47 +87,12 @@
    "source": [
     "print('Non-framerate adjusted renditions:')\n",
     "df_non_fps = df.query('not rendition_type.str.contains(\"fps\")')\n",
-    "print('Timestamp-based frame selection:')\n",
-    "print(classification_report(df_non_fps['ground_truth'], df_non_fps['prediction']))\n",
-    "print('Index-based frame selection:')\n",
-    "print(classification_report(df_non_fps['ground_truth'], df_non_fps['prediction_old']))"
+    "print(classification_report(df_non_fps['ground_truth'], df_non_fps['prediction']))"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 22,
-   "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "Non-matching Index-based and TS-based predictions\n",
-      "                                is_correct     score  score_old  time_sec  \\\n",
-      "master_filename rendition_type                                              \n",
-      "u05F29jXYNE.mp4 720p                  True -1.468984   0.341502  2.425891   \n",
-      "                720p_watermark       False -1.687979   0.579660  2.531694   \n",
-      "\n",
-      "                                time_sec_old ground_truth prediction  \\\n",
-      "master_filename rendition_type                                         \n",
-      "u05F29jXYNE.mp4 720p                1.834705      Correct     Tamper   \n",
-      "                720p_watermark      1.904839       Tamper     Tamper   \n",
-      "\n",
-      "                               prediction_old  \n",
-      "master_filename rendition_type                 \n",
-      "u05F29jXYNE.mp4 720p                  Correct  \n",
-      "                720p_watermark        Correct  \n"
-     ]
-    }
-   ],
-   "source": [
-    "print('Non-matching Index-based and TS-based predictions')\n",
-    "print(df_non_fps[df_non_fps['prediction']!=df_non_fps['prediction_old']])"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 23,
+   "execution_count": 103,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -147,21 +103,21 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "FPS-adjusted renditions report:\n",
+      "All FPS-adjusted renditions report:\n",
       "              precision    recall  f1-score   support\n",
       "\n",
-      "     Correct       0.91      0.59      0.72        88\n",
-      "      Tamper       0.49      0.87      0.62        39\n",
+      "     Correct       0.85      0.72      0.78        88\n",
+      "      Tamper       0.75      0.88      0.81        88\n",
       "\n",
-      "    accuracy                           0.68       127\n",
-      "   macro avg       0.70      0.73      0.67       127\n",
-      "weighted avg       0.78      0.68      0.69       127\n",
+      "    accuracy                           0.80       176\n",
+      "   macro avg       0.80      0.80      0.79       176\n",
+      "weighted avg       0.80      0.80      0.79       176\n",
       "\n"
      ]
     }
    ],
    "source": [
-    "print('FPS-adjusted renditions report:')\n",
+    "print('All FPS-adjusted renditions report:')\n",
     "df_fps = df.query('rendition_type.str.contains(\"fps\")')\n",
     "df_fps\n",
     "print(classification_report(df_fps['ground_truth'], df_fps['prediction']))"
@@ -169,7 +125,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 24,
+   "execution_count": 99,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -184,12 +140,12 @@
       "FPS-adjusted 30-60 FPS renditions report:\n",
       "              precision    recall  f1-score   support\n",
       "\n",
-      "     Correct       0.87      0.69      0.77        39\n",
-      "      Tamper       0.59      0.81      0.68        21\n",
+      "     Correct       0.81      0.90      0.85        39\n",
+      "      Tamper       0.89      0.79      0.84        39\n",
       "\n",
-      "    accuracy                           0.73        60\n",
-      "   macro avg       0.73      0.75      0.73        60\n",
-      "weighted avg       0.77      0.73      0.74        60\n",
+      "    accuracy                           0.85        78\n",
+      "   macro avg       0.85      0.85      0.85        78\n",
+      "weighted avg       0.85      0.85      0.85        78\n",
       "\n"
      ]
     }
@@ -197,9 +153,43 @@
    "source": [
     "print('FPS-adjusted 30-60 FPS renditions report:')\n",
     "df_fps = df.query('rendition_type.str.contains(\"30-60fps\") or rendition_type.str.contains(\"60-30fps\")')\n",
-    "df_fps\n",
     "print(classification_report(df_fps['ground_truth'], df_fps['prediction']))"
    ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 102,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "Non-divisible FPS-adjusted renditions report:\n",
+      "              precision    recall  f1-score   support\n",
+      "\n",
+      "     Correct       0.79      0.78      0.78       113\n",
+      "      Tamper       0.78      0.79      0.78       113\n",
+      "\n",
+      "    accuracy                           0.78       226\n",
+      "   macro avg       0.78      0.78      0.78       226\n",
+      "weighted avg       0.78      0.78      0.78       226\n",
+      "\n"
+     ]
+    }
+   ],
+   "source": [
+    "print('Non-divisible FPS-adjusted renditions report:')\n",
+    "df_fps = df.query('not rendition_type.str.contains(\"30-60fps\") and not rendition_type.str.contains(\"60-30fps\")')\n",
+    "print(classification_report(df_fps['ground_truth'], df_fps['prediction']))"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": []
   }
  ],
  "metadata": {