--- a/feature_engineering/notebooks/Fps_renditions_gpu.ipynb
+++ b/feature_engineering/notebooks/Fps_renditions_gpu.ipynb
@@ -8,13 +8,13 @@
     }
    },
    "source": [
-    "# Classification metrics for FPS-adjusted renditions\n",
+    "# Classification metrics for FPS-adjusted renditions generated with GPU codec\n",
     "#### Data is generated with testing/test_var_fps_classification.py script"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 44,
+   "execution_count": 5,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -28,7 +28,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 45,
+   "execution_count": 6,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -37,14 +37,14 @@
    "outputs": [],
    "source": [
     "# load data\n",
-    "df = pd.read_csv('test_fps_renditions.csv')\n",
+    "df = pd.read_csv('test_fps_renditions_gpu.csv')\n",
     "df.set_index(['master_filename', 'rendition_type'], inplace=True)\n",
     "df.sort_index(inplace=True)"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 46,
+   "execution_count": 7,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -59,7 +59,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 47,
+   "execution_count": 8,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -92,7 +92,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 48,
+   "execution_count": 10,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -106,12 +106,12 @@
       "All FPS-adjusted renditions report:\n",
       "              precision    recall  f1-score   support\n",
       "\n",
-      "     Correct       0.85      0.72      0.78        88\n",
-      "      Tamper       0.75      0.88      0.81        88\n",
+      "     Correct       0.82      0.52      0.64        44\n",
+      "      Tamper       0.65      0.89      0.75        44\n",
       "\n",
-      "    accuracy                           0.80       176\n",
-      "   macro avg       0.80      0.80      0.79       176\n",
-      "weighted avg       0.80      0.80      0.79       176\n",
+      "    accuracy                           0.70        88\n",
+      "   macro avg       0.74      0.70      0.69        88\n",
+      "weighted avg       0.74      0.70      0.69        88\n",
       "\n"
      ]
     }
@@ -125,7 +125,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 49,
+   "execution_count": 11,
    "metadata": {
     "pycharm": {
      "name": "#%%\n"
@@ -140,12 +140,12 @@
       "FPS-adjusted 30-60 FPS renditions report:\n",
       "              precision    recall  f1-score   support\n",
       "\n",
-      "     Correct       0.81      0.90      0.85        39\n",
-      "      Tamper       0.89      0.79      0.84        39\n",
+      "     Correct       0.80      0.51      0.62        39\n",
+      "      Tamper       0.64      0.87      0.74        39\n",
       "\n",
-      "    accuracy                           0.85        78\n",
-      "   macro avg       0.85      0.85      0.85        78\n",
-      "weighted avg       0.85      0.85      0.85        78\n",
+      "    accuracy                           0.69        78\n",
+      "   macro avg       0.72      0.69      0.68        78\n",
+      "weighted avg       0.72      0.69      0.68        78\n",
       "\n"
      ]
     }
@@ -158,7 +158,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 50,
+   "execution_count": 12,
    "metadata": {},
    "outputs": [
     {
@@ -168,12 +168,12 @@
       "Non-divisible FPS-adjusted renditions report:\n",
       "              precision    recall  f1-score   support\n",
       "\n",
-      "     Correct       0.79      0.78      0.78       113\n",
-      "      Tamper       0.78      0.79      0.78       113\n",
+      "     Correct       0.75      0.91      0.82        69\n",
+      "      Tamper       0.89      0.70      0.78        69\n",
       "\n",
-      "    accuracy                           0.78       226\n",
-      "   macro avg       0.78      0.78      0.78       226\n",
-      "weighted avg       0.78      0.78      0.78       226\n",
+      "    accuracy                           0.80       138\n",
+      "   macro avg       0.82      0.80      0.80       138\n",
+      "weighted avg       0.82      0.80      0.80       138\n",
       "\n"
      ]
     }
@@ -186,26 +186,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
-   "outputs": [],
-   "source": [],
-   "metadata": {
-    "collapsed": false,
-    "pycharm": {
-     "name": "#%%\n"
-    }
-   }
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "metadata": {},
-   "outputs": [],
-   "source": []
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 7,
    "metadata": {},
    "outputs": [],
    "source": []