--- a/feature_engineering/notebooks/Tools.ipynb
+++ b/feature_engineering/notebooks/Tools.ipynb
@@ -62,7 +62,7 @@
     "# 1.-Rendition creation\n",
     "## 1.0.- Scaling and bitrate reduction\n",
     "\n",
-    "In first place, a dataset with original videos is required. We propose the collection provided in the YT8M notebook available [here](https://github.com/epiclabs-io/YT8M/blob/master/yt8m-crawler.ipynb). This would provide a nice collection of renditions at different resolutions and with adequate bitrates. Nevertheless, feel free to supply your own set of videos and renditions at whatever bitrate you consider.\n",
+    "In first place, a dataset with original videos is required. We propose the collection provided in the YT8M notebook available [here](https://github.com/epiclabs-io/YT8M/blob/master/yt8m-crawler.ipynb). Once CSV file with the videos metadata is generated, use [downloader](../../YT8M_downloader/downloader.py) script to download videos. Consider using various format filter values to get e.g. 60 fps videos.\n",
     "\n",
     "## 1.1.- Watermarks\n",
     "\n",
@@ -428,6 +428,49 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
+    "## 1.11.- FPS converter\n",
+    "\n",
+    "This script converts source videos to common frame rates, and puts results into corresponding folders. It can be applied after all other renditions are created.\n",
+    "- The input path (-i or --input) which is the folder containing the renditions.\n",
+    "- The output path (-o or --output) which is the folder where the renditions are going to be stored.\n",
+    "- GPU flag (-g or --gpu). Whether to use GPU codecs. You need to have Ffmpeg built with Nvidia Codecs for it to work. Default is false.\n",
+    "\n",
+    "The output folder names will have source-target FPS suffix appended.\n",
+    "\n",
+    "```\n",
+    "├── 1080p_watermark_60-24fps\n",
+    "│   ├── <filename>.mp4\n",
+    "├── 720p_watermark_60-24fps\n",
+    "│   ├── <filename>.mp4\n",
+    "├── 480p_watermark_60-24fps\n",
+    "│   ├── <filename>.mp4\n",
+    "├── 360p_watermark_60-24fps\n",
+    "│   ├── <filename>.mp4\n",
+    "├── 240p_watermark_60-24fps\n",
+    "│   ├── <filename>.mp4\n",
+    "├── 144p_watermark_60-24fps\n",
+    "│   ├── <filename>.mp4\n",
+    "```"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "outputs": [],
+   "source": [
+    "output_path = (current_path / \"../data\").resolve()\n",
+    "%run -i '/scripts/fps_converter.py' -i $input_path -o $output_path"
+   ],
+   "metadata": {
+    "collapsed": false,
+    "pycharm": {
+     "name": "#%%\n"
+    }
+   }
+  },
+  {
+   "cell_type": "markdown",
+   "source": [
     "# 2 Metric extraction\n",
     "\n",
     "In order to extract metrics from the videos above and being able to compare them, a set of scripts are made available below.\n",
@@ -467,7 +510,10 @@
     "0.985460, 32.521368, 43.338460, 42.580399\n",
     "0.985896, 32.670122, 43.325404, 42.529248\n",
     "```"
-   ]
+   ],
+   "metadata": {
+    "collapsed": false
+   }
   },
   {
    "cell_type": "code",