--- a/scripts/asset_processor/__init__.py
+++ b/scripts/asset_processor/__init__.py
@@ -1,0 +1,3 @@
+from .video_asset_processor import *
+from .video_capture import *
+from .video_metrics import *