--- a/verifier/__init__.py
+++ b/verifier/__init__.py
@@ -1,0 +1,1 @@
+from verifier import verifier