… README.md
… feature_engineering/notebooks/test_fps_renditions.csv
… machine_learning/output/models/CB_Full_v2.cbm
… machine_learning/output/models/OCSVM.pickle.dat
… machine_learning/output/models/UL_StandardScaler.pickle.dat
… machine_learning/output/models/param_OCSVM.json
… requirements.txt
… scripts/README.md
… scripts/asset_processor/video_asset_processor_opencv.py
… scripts/asset_processor/video_capture.py
… scripts/fps_converter.py
… scripts/watermark.py
… testing/mini_dataset_gen.py
… testing/test_var_fps.py
… testing/test_var_fps_classification.py
… testing/tests/conftest.py
… testing/tests/data/0fIdY5IAnhY_60.mp4
… testing/tests/test_var_fps.py
… verifier/verifier.py