These changes should resolve issues #93 #105.

# Theory
When renditions frame rates are adjusted compared to source (master) video, it's no longer possible to use frame indexes to establish frame correspondence when calculating metrics. In this case, frames should be matched based on presentation timestamps (PTS). There are two possible scenarios:
1. Master and rendition frame rates are divisible, and master frame rate is lower (e.g. 30 FPS to 60 FPS)
In this case, rendition frame will always exist at timestamp samples from master video.
2. All other cases (e.g. 60 to 30 FPS, 30 to 24 FPS, ...)
Such cases require calculating differences between arbitrary number of master timestamps and candidate rendition timestamps to pick frames which were preserved from master at closest possible positions during transcoding.

# Implementation
1. PTS extraction is based on parsing debug output of Ffmpeg CLI. It is run for each video with null output.
2. If number of samples is set to a specific number, an increased number of candidate frames is picked from master video, based on master-rendition frame rate ratio. Then, for each rendition, closest possible frames are chosen from these candidate frames for matching. This makes matching process somewhat probabilistic, but doesn't require storing all master frames or scanning master video once for each rendition.
3. Debug output option is added to VideoAssetProcessor. When enabled, it dumps frame pairs as PNG images to assist in visually spotting frame differences.
4. Classification accuracy is tested on a subset of YT8K dataset with frame rate - adjusted renditions generated by Ffmpeg complex filter. See [notebook](feature_engineering/notebooks/Fps_renditions.ipynb).

# Potential issues
1. Each video is decoded twice - with Ffmpeg to extract timestamps, and then with OpenCV to read frames. Tests are showing 28% performance decrease compared to OpenCV-only video decoding. 
2. Memory requirements are increased, because algorithm is caching more master frames for a given number of samples requested.
3. If timestamps are heavily manipulated during transcoding, it's not possible to get matching frames with this approach.
4. Code is tested only with Ffmpeg 4.2.1. Debug output format may be different in other versions and timestamp reading may not work.