@yondonfu 2020-06-16
Can the requirements.txt file be updated so that it installs OpenCV from the custom binary wheel (https://github.com/cyberj0g/OpenCV-Custom)?

Also, it would be helpful if the custom OpenCV source was available so that the commits with the updates for returning the raw PTS from ffmpeg for `CAP_PROP_POS_MSEC` (instead of the frame number multiplied by the FPS) are visible. If there is a forked OpenCV repo with these changes the custom binary wheel files could be kept there as well.