@cyberj0g 2020-06-16
Unfortunately, it seems requirements.txt can't contain a direct binary wheel link, and I have created the wheel by modifying original opencv-python wheel. It seems, the option which would allow to install custom OpenCV from requirements.txt (without the need to build it locally on install using setup.py of original opencv-python) is publishing this version to PyPI under other name.